feat: handleToggleTheme, refactor code
