+++ packages/wiki-rev-diff/src/utils.ts
@@ -1,0 +1,6 @@
+export function getCookie(name: string): string | undefined {
+  const value = '; ' + document.cookie;
+  const parts = value.split('; ' + name + '=');
+  if (parts.length === 2) return parts.pop()?.split(';').shift();
+  return undefined;
+}
